did a bunch of role stuff and UI functionality, save drawing callback missing 144th led value

+++ app/assets/javascripts/ellie_dee.js
@@ -1,4 +1,5 @@
 $(document).ready(function() {
+  var userRole = $("#userRole").val();
   buildMatrix();
 
   $(function () {
@@ -13,27 +14,37 @@ $(document).ready(function() {
     });
   });
 
+  $("#accessRole").popover({
+    trigger: "hover",
+    placement: 'bottom',
+    title: 'Access Roles',
+    html: true,
+    content: $('#accessInfo').html()
+  });
+
   $(".led").on("click", function() {
 
     $(this).css('background-color', $("#currentColor").val());
     var currentColors = getColors();
 
-    $.ajax({
-      dataType: 'json',
-      url: '/drawings/' + $("#ellieDeeId").val().toString() + '.json',
-      method: 'PUT',
-      contentType: "application/json",
-      data: JSON.stringify({
-        "leds": currentColors
-      }),
-      success: function(data) {
-        console.log("EllieDee Successfully Updated: " + JSON.stringify(data));
-        setColors(data.leds);
-      },
-      error: function(jqXHR, textStatus, error) {
-        console.log("EllieDee Failed to Update: " + error);
-      }
-    });
+    if (userRole === 'admin' || userRole === 'ctrl') {
+      $.ajax({
+        dataType: 'json',
+        url: '/drawings/' + $("#ellieDeeId").val().toString() + '.json',
+        method: 'PUT',
+        contentType: "application/json",
+        data: JSON.stringify({
+          "leds": currentColors
+        }),
+        success: function(data) {
+          console.log("EllieDee Successfully Updated: " + JSON.stringify(data));
+          setColors(data.leds);
+        },
+        error: function(jqXHR, textStatus, error) {
+          console.log("EllieDee Failed to Update: " + error);
+        }
+      });
+    };
   });
 
   $("#ledOff").on("click", function() {
@@ -52,43 +63,64 @@ $(document).ready(function() {
   }).on('click', function() {
     $("#saveName").focus();
     $("#save").on("click", function() {
-      var saveName = $("#saveName").val();
-      console.log("Save Name: ", saveName);
-      var saveLeds = getColors();
-      saveLeds = "[" + saveLeds.toString() + "]";
       var saveUserId = $("#saveUserId").val();
-      console.log("User ID: ", saveUserId);
+      var saveName = $("#saveName").val();
+      var saveLeds = getColors();
+      console.log("Save Colors Array Length: " + saveLeds.length);
+      saveLeds = "[" + saveLeds.toString() + "]";
 
-      $.ajax({
-        dataType: 'json',
-        url: '/drawings/',
-        method: 'POST',
-        contentType: "application/json",
-        data: JSON.stringify({
-          "name": saveName,
-          "leds": saveLeds,
-          "user_id": saveUserId
-        }),
-        success: function(data) {
-          console.log(saveDrawing + " Successfully Saved: " + data);
-          $("#saveDrawing").popover('hide')
-          $("#userDrawings").append(
-            "<tr><td>" + data.name + "</td><td><i id='d" + data.user_id + "' class='fa fa-play-circle'></i></td><td><a data-confirm='Are you sure?' rel='nofollow' data-method='delete' href='/drawings/" + data.id + "' <i class='fa fa-trash'></i></a></td></tr>"
-          );
-        },
-        error: function(jqXHR, textStatus, error) {
-          console.log(saveDrawing + " Failed to Save: " + error);
-          if (error === "Unprocessable Entity") {
-            $("#saveAlert").text("Uh oh, looks like you're already using this name. Pick a new name or try again after deleting the old one below!").attr("type", "text");
-          };
-        }
-      });
+      if (userRole === 'admin' || userRole === 'ctrl' || userRole === 'reader') {
+        $.ajax({
+          dataType: 'json',
+          url: '/drawings/',
+          method: 'POST',
+          contentType: "application/json",
+          data: JSON.stringify({
+            "name": saveName,
+            "leds": saveLeds,
+            "user_id": saveUserId
+          }),
+          success: function(data) {
+            // console.log(saveDrawing + " Successfully Saved: " + data);
+            console.log("Callback Array Length: " + data.leds.length);
+            $("#saveDrawing").popover('hide')
+            $("#userDrawings").append(
+              "<tr><td>" + data.name + "</td><td><i id='d" + data.user_id + "' class='fa fa-play-circle'></i></td><td><a data-confirm='Are you sure?' rel='nofollow' data-method='delete' href='/drawings/" + data.id + "' <i class='fa fa-trash'></i></a></td></tr>"
+            );
+          },
+          error: function(jqXHR, textStatus, error) {
+            console.log(saveDrawing + " Failed to Save: " + error);
+            if (error === "Unprocessable Entity") {
+              $("#saveAlert").text("Uh oh, looks like you're already using this name. Pick a new name or try again after deleting the old one below!").attr("type", "text");
+            };
+          }
+        });
+      };
     });
   });
 
+  $(".fa-play-circle").on("click", function() {
+    var drawingId = $(this).attr('id');
+    drawingId = drawingId.slice(1, drawingId.length);
+
+    $.ajax({
+      dataType: 'json',
+      url: '/drawings/' + drawingId + '.json',
+      method: 'GET',
+      contentType: "application/json",
+      success: function(data) {
+        console.log("GET drawing was Successful. LEDs: " + data.leds);
+        console.log("LED Array Length: " + data.leds.length);
+        setColors(data.leds);
+      },
+      error: function(jqXHR, textStatus, error) {
+        console.log("GET drawing Failed: " + error);
+      }
+    });
+  });
 });
 
-//Creates a function that generates a really long string with entire table HTML and sends AJAX call to reset LED array in database
+//Function to generate string for entire table HTML and sends AJAX call to reset LED array in db
 function buildMatrix() {
   var matrixString = "";
   var ledNum = 143;
@@ -104,44 +136,49 @@ function buildMatrix() {
   resetEllieDee();
 }
 
+//Reset drawing to default/cleared state
 function resetEllieDee() {
   var blankArray = [];
   for (var i = 0; i < 145; i++) {
     blankArray.push("#808080");
   }
-  blankArray = "[" + blankArray.toString() + "]";
 
-  $.ajax({
-    dataType: 'json',
-    url: '/drawings/' + $("#ellieDeeId").val().toString() + '.json',
-    method: 'PUT',
-    contentType: "application/json",
-    data: JSON.stringify({
-      "leds": blankArray
-    }),
-    success: function(data) {
-      console.log("EllieDee Successfully Reset. LEDs: " + data.leds);
-      setColors(data.leds);
-    },
-    error: function(jqXHR, textStatus, error) {
-      console.log("EllieDee Failed to Reset: " + error);
-    }
-  });
+  if (userRole === 'admin' || userRole === 'ctrl') {
+    blankArray = "[" + blankArray.toString() + "]";
+    $.ajax({
+      dataType: 'json',
+      url: '/drawings/' + $("#ellieDeeId").val().toString() + '.json',
+      method: 'PUT',
+      contentType: "application/json",
+      data: JSON.stringify({
+        "leds": blankArray
+      }),
+      success: function(data) {
+        console.log("EllieDee Successfully Reset. LEDs: " + data.leds);
+        setColors(data.leds);
+      },
+      error: function(jqXHR, textStatus, error) {
+        console.log("EllieDee Failed to Reset: " + error);
+      }
+    });
+  } else {
+    setColors(blankArray);
+  };
 }
 
+//Iterate through the drawing to get the current state of all leds
 function getColors() {
   var colors = [];
 
   for (var i = 0; i <= 143; i++) {
     var currentLed = "#" + i.toString();
     var ledColor = rgb2hex($(currentLed).css('backgroundColor'));
-    // console.log("getColor LED: ", ledColor);
     colors.push(ledColor);
   }
-  console.log("Colors: ", colors);
   return colors;
 }
 
+//Set all colors based on response from db or saved drawing
 function setColors(array) {
   for (var i = 0; i < array.length; i++) {
     var currentLed = "#" + i.toString();
